feat: Separated network infra installers.
