+++ Assets/Conekton/Multiplayer/Scripts/Installers/StubNetworkInfraInstaller.cs
@@ -1,0 +1,18 @@
+using UnityEngine;
+using Zenject;
+
+using Conekton.ARMultiplayer.Avatar.Domain;
+using Conekton.ARMultiplayer.NetworkMultiplayer.Domain;
+using Conekton.ARMultiplayer.NetworkMultiplayer.Infrastructure;
+
+namespace Conekton.ARMultiplayer.NetworkMultiplayer.Application
+{
+    public class StubNetworkInfraInstaller : MonoInstaller
+    {
+        public override void InstallBindings()
+        {
+            Container.Bind<IMultiplayerNetworkInfrastructure>().To<StubNetworkInfra>().AsCached();
+        }
+    }
+}
+
